volunteer

+++ app/Http/Controllers/ImageGalleryController.php
@@ -16,9 +16,13 @@ class ImageGalleryController extends Controller
      */
     public function index()
     {
-        return view('backend.pages.media.image_gallery.index',[
-            'imageGalleries' => ImageGallery::latest()->paginate(30),
-        ]);
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.image_gallery.index',[
+                'imageGalleries' => ImageGallery::latest()->paginate(30),
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -28,7 +32,11 @@ class ImageGalleryController extends Controller
      */
     public function create()
     {
-        return view('backend.pages.media.image_gallery.create');
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.image_gallery.create');
+        }else{
+            return abort(404);
+        }
     }
     /**
      * Store a newly created resource in storage.
@@ -38,40 +46,44 @@ class ImageGalleryController extends Controller
      */
     public function store(Request $request)
     {
-        if($request->hasFile('images')){
-            $size = 0;
-            foreach ($request->file('images') as $image) {
-                $ext = Str::lower($image->getClientOriginalExtension());
-                $valid_ext = ['jpg','jpeg','png','webp'];
-                if(!in_array($ext, $valid_ext)){
-                    return back()->with('validation_error','Uploaded file only suport jpg, jpeg, png or webp');
+        if(auth()->user()->can('media management')){
+            if($request->hasFile('images')){
+                $size = 0;
+                foreach ($request->file('images') as $image) {
+                    $ext = Str::lower($image->getClientOriginalExtension());
+                    $valid_ext = ['jpg','jpeg','png','webp'];
+                    if(!in_array($ext, $valid_ext)){
+                        return back()->with('validation_error','Uploaded file only suport jpg, jpeg, png or webp');
+                    }
+                    $size += $image->getSize();
                 }
-                $size += $image->getSize();
-            }
 
-            if($size < 10485760){
-                $loopKey = 0;
-                foreach ($request->file('images') as $key => $image) {
-                    $loopKey += $key;
-                    $ImageGallery = new ImageGallery();
-                    $newName = 'Image-gallery'.Str::random(5).time().'.'.$image->getClientOriginalExtension();
-                    $path = public_path('images/media/image_gallery/'.$newName);
-                    Image::make($image)->save($path);
-                    $ImageGallery->alt_text =  pathinfo($image->getClientOriginalname(),PATHINFO_FILENAME);
-                    $ImageGallery->image = $newName;
-                    $ImageGallery->save();
-                }
-                if($loopKey > 0){
-                    return redirect()->route('image-gallery.index')->with('success','New Images Added!');
+                if($size < 10485760){
+                    $loopKey = 0;
+                    foreach ($request->file('images') as $key => $image) {
+                        $loopKey += $key;
+                        $ImageGallery = new ImageGallery();
+                        $newName = 'Image-gallery'.Str::random(5).time().'.'.$image->getClientOriginalExtension();
+                        $path = public_path('images/media/image_gallery/'.$newName);
+                        Image::make($image)->save($path);
+                        $ImageGallery->alt_text =  pathinfo($image->getClientOriginalname(),PATHINFO_FILENAME);
+                        $ImageGallery->image = $newName;
+                        $ImageGallery->save();
+                    }
+                    if($loopKey > 0){
+                        return redirect()->route('image-gallery.index')->with('success','New Images Added!');
+                    }else{
+                        return redirect()->route('image-gallery.index')->with('success','New Image Added!');
+                    }
                 }else{
-                    return redirect()->route('image-gallery.index')->with('success','New Image Added!');
+                    return back()->with('validation_error','Uploaded file can not longer than 10MB');
                 }
-            }else{
-                return back()->with('validation_error','Uploaded file can not longer than 10MB');
-            }
 
+            }else{
+                return back()->with('validation_error','Please choose minimum one image to upload');
+            }
         }else{
-            return back()->with('validation_error','Please choose minimum one image to upload');
+            return abort(404);
         }
     }
 
@@ -120,38 +132,54 @@ class ImageGalleryController extends Controller
         //
     }
     public function getImage($id){
-        $img = ImageGallery::find($id);
-        return response()->json($img);
+        if(auth()->user()->can('media management')){
+            $img = ImageGallery::find($id);
+            return response()->json($img);
+        }else{
+            return abort(404);
+        }
     }
     public function updateImage(Request $request){
-        $image = ImageGallery::find($request->image_id);
-        $image->alt_text = $request->alt_name;
-        if($image->save()){
-            return back()->with('success','Alt Name Saved');
+        if(auth()->user()->can('media management')){
+            $image = ImageGallery::find($request->image_id);
+            $image->alt_text = $request->alt_name;
+            if($image->save()){
+                return back()->with('success','Alt Name Saved');
+            }else{
+                return back()->with('error','Failed');
+            }
         }else{
-            return back()->with('error','Failed');
+            return abort(404);
         }
     }
     public function updateTrash($id){
-        $image = ImageGallery::find($id);
-        if($image->delete()){
-            return back()->with('success','Image Moved to trash');
-        }else{
-            return back()->with('error','Failed');
-        }
-    }
-    public function updateDeletePermanently($id){
-        $image = ImageGallery::find($id);
-        $oldImage = public_path('images/media/image_gallery/'.$image->image);
-        if($oldImage){
-            unlink($oldImage);
-            if($image->forceDelete()){
+        if(auth()->user()->can('media management')){
+            $image = ImageGallery::find($id);
+            if($image->delete()){
                 return back()->with('success','Image Moved to trash');
             }else{
                 return back()->with('error','Failed');
             }
         }else{
-            return back()->with('error','Failed');
+            return abort(404);
+        }
+    }
+    public function updateDeletePermanently($id){
+        if(auth()->user()->can('media management')){
+            $image = ImageGallery::find($id);
+            $oldImage = public_path('images/media/image_gallery/'.$image->image);
+            if($oldImage){
+                unlink($oldImage);
+                if($image->forceDelete()){
+                    return back()->with('success','Image Moved to trash');
+                }else{
+                    return back()->with('error','Failed');
+                }
+            }else{
+                return back()->with('error','Failed');
+            }
+        }else{
+            return abort(404);
         }
     }
 }
