+++ app/Http/Controllers/PublicationsController.php
@@ -18,9 +18,13 @@ class PublicationsController extends Controller
      */
     public function index()
     {
-        return view('backend.pages.media.publications.index',[
-            'publications' => Publications::latest()->paginate(12),
-        ]);
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.publications.index',[
+                'publications' => Publications::latest()->paginate(12),
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -30,9 +34,13 @@ class PublicationsController extends Controller
      */
     public function create()
     {
-        return view('backend.pages.media.publications.create',[
-            'projects' => Project::all(),
-        ]);
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.publications.create',[
+                'projects' => Project::all(),
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -43,33 +51,37 @@ class PublicationsController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'headline' => "required",
-            'url' => "required",
-            'media' => "required",
-            'published_date' => "required",
-            'featured_image' => "required",
-        ]);
-        $publication = new Publications;
-        $publication->headline = $request->headline;
-        $publication->slug = Str::slug($request->headline);
-        $publication->url = $request->url;
-        $publication->project_id = $request->project_id;
-        $publication->media = $request->media;
-        $publication->published_date = $request->published_date;
-        $publication->save();
-        if($request->hasFile('featured_image')){
-
-            $image = $request->file('featured_image');
-            // return $image;
-            $newName = Str::slug($request->headline).'-'.Str::random(5).'.'.$image->getClientOriginalExtension();
-            $destination = public_path('images/media/publications/'.$newName);
-            // return 'hello';
-            Image::make($image)->save($destination);
-            $publication->featured_image = $newName;
+        if(auth()->user()->can('media management')){
+            $request->validate([
+                'headline' => "required",
+                'url' => "required",
+                'media' => "required",
+                'published_date' => "required",
+                'featured_image' => "required",
+            ]);
+            $publication = new Publications;
+            $publication->headline = $request->headline;
+            $publication->slug = Str::slug($request->headline);
+            $publication->url = $request->url;
+            $publication->project_id = $request->project_id;
+            $publication->media = $request->media;
+            $publication->published_date = $request->published_date;
             $publication->save();
+            if($request->hasFile('featured_image')){
+
+                $image = $request->file('featured_image');
+                // return $image;
+                $newName = Str::slug($request->headline).'-'.Str::random(5).'.'.$image->getClientOriginalExtension();
+                $destination = public_path('images/media/publications/'.$newName);
+                // return 'hello';
+                Image::make($image)->save($destination);
+                $publication->featured_image = $newName;
+                $publication->save();
+            }
+            return redirect()->route('publications.index')->with('success','Add new publication');
+        }else{
+            return abort(404);
         }
-        return redirect()->route('publications.index')->with('success','Add new publication');
     }
 
     /**
@@ -80,9 +92,13 @@ class PublicationsController extends Controller
      */
     public function show($id)
     {
-        return view('backend.pages.media.publications.show',[
-            'publication' => Publications::find($id),
-        ]);
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.publications.show',[
+                'publication' => Publications::find($id),
+            ]);
+        }else{
+            return abort(404);
+        }
 
 
     }
@@ -95,10 +111,13 @@ class PublicationsController extends Controller
      */
     public function edit($id)
     {
-
-        return view('backend.pages.media.publications.edit',[
-            'publication' => Publications::find($id),
-        ]);
+        if(auth()->user()->can('media management')){
+            return view('backend.pages.media.publications.edit',[
+                'publication' => Publications::find($id),
+            ]);
+        }else{
+            return abort(404);
+        }
     }
 
     /**
@@ -110,36 +129,40 @@ class PublicationsController extends Controller
      */
     public function update(Request $request)
     {
-        $request->validate([
-            'headline' => "required",
-            'url' => "required",
-            'media' => "required",
-            'published_date' => "required",
-        ]);
+        if(auth()->user()->can('media management')){
+            $request->validate([
+                'headline' => "required",
+                'url' => "required",
+                'media' => "required",
+                'published_date' => "required",
+            ]);
 
-        $publication = Publications::find($request->publication_id);
-        $publication->headline = $request->headline;
-        $publication->slug = Str::slug($request->headline);
-        $publication->url = $request->url;
-        $publication->project_id = $request->project_id;
-        $publication->media = $request->media;
-        $publication->published_date = $request->published_date;
-        $publication->save();
-
-        if($request->hasFile('featured_image')){
-
-            $oldImage = public_path('images/media/publications/'.$publication->featured_image);
-            if(file_exists($oldImage)){
-                unlink($oldImage);
-            }
-            $image = $request->file('featured_image');
-            $newName = Str::slug($request->headline).'-'.Str::random(5).'.'.$image->getClientOriginalExtension();
-            $destination = public_path('images/media/publications/'.$newName);
-            Image::make($image)->save($destination);
-            $publication->featured_image = $newName;
+            $publication = Publications::find($request->publication_id);
+            $publication->headline = $request->headline;
+            $publication->slug = Str::slug($request->headline);
+            $publication->url = $request->url;
+            $publication->project_id = $request->project_id;
+            $publication->media = $request->media;
+            $publication->published_date = $request->published_date;
             $publication->save();
+
+            if($request->hasFile('featured_image')){
+
+                $oldImage = public_path('images/media/publications/'.$publication->featured_image);
+                if(file_exists($oldImage)){
+                    unlink($oldImage);
+                }
+                $image = $request->file('featured_image');
+                $newName = Str::slug($request->headline).'-'.Str::random(5).'.'.$image->getClientOriginalExtension();
+                $destination = public_path('images/media/publications/'.$newName);
+                Image::make($image)->save($destination);
+                $publication->featured_image = $newName;
+                $publication->save();
+            }
+            return redirect()->route('publications.index')->with('success','Add new publication');
+        }else{
+            return abort(404);
         }
-        return redirect()->route('publications.index')->with('success','Add new publication');
     }
 
     /**
@@ -150,13 +173,17 @@ class PublicationsController extends Controller
      */
     public function destroy($id)
     {
-        $publication = Publications::find($id);
-        if($publication->featured_image){
-            if(file_exists(public_path("images/media/publications/".$publication->featured_image))){
-                unlink(public_path("images/media/publications/".$publication->featured_image));
+        if(auth()->user()->can('media management')){
+            $publication = Publications::find($id);
+            if($publication->featured_image){
+                if(file_exists(public_path("images/media/publications/".$publication->featured_image))){
+                    unlink(public_path("images/media/publications/".$publication->featured_image));
+                }
             }
+            $publication->delete();
+            return back()->with('success','Publication Deleted!');
+        }else{
+            return abort(404);
         }
-        $publication->delete();
-        return back()->with('success','Publication Deleted!');
     }
 }
